myproject/requirements.txt
Django==1.11.10
djangorestframework==3.7.7
pkg-resources==0.0.0
pytz==2018.3
